+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - embedded objects that fail to load should now be visible as error message, even if they provide no alternatives
 - bugfix: Row filter works again if no columns are merged.
 - bugfix: Swapping rows with invisible rows should work while column merging is active.
+- bugfix: Resetting the column order will preserve visibilities of columns
 
 ## 0.6.1